Workaround for bug in PlotWidget.__getattr__ in pg v0.10
Pyqtgraph v0.10.0 has a bug in PlotWidget: the getattr implementation
raises a NameError instead of an AttributeError. This makes hasattr()
fail. As a workaround, fix it in tpg's TaurusPlot and TaurusTrend.

Note: since the bug is already fixed in the develop branch (unreleased)
of pyqtgraph, this commit can be reverted afer the next release of
pyqtgraph (v>0.10.0).