Make requirements more explicit
The taurus[taurus-qt] dependency includes guiqwt (and guidata from it).
The guidata and guiqwt dependencies ar a PITA to install using pip
(guidata fails to install unless a cython and sip are installed,
and guiqwt fails to install unless guidata is installed).
But we do not really need guidata or guiqwt for taurus_pyqtgraph, so we
can work around the issues and have a cleaner environment and
configuration by not using  taurus[taurus-qt] dependency and instead
depending explicitly on ply and lxml.

Also pin the corresponding dependencies in requirements.txt to get more
reproducible environments.