Fix issue in plots & trend constructors with Qt4
TaurusPlot and TaurusTrend fail to instantiate when using PyQt5 and
python3. Commit 74c7c9d7 intended to fix this same issue, but only
fixed it for python2. In fact it was a wrong fix because it checked
against the python version when it should have checked against Qt
version (the issue is only Qt version dependent). Fix it by checking
against QtCore.QT_VERSION