Improve the forced read tool
- Change the forced read tool from a widget action with an spinbox to a
QAction that triggers a dialog. This avoids some weird UI glitches
in which the menu disappears when clicking the spin box for edit and
it also prevents from applying while editing (Fixes #45)

- Enable the forced reading also on newly added trendsets. Allow
the buffer tool to react on newly added trend sets and apply the
reading period to them. This requires
https://github.com/pyqtgraph/pyqtgraph/pull/1388 to be integrated
in pyqtgraph.